// worklenz-frontend/src/app/administrator/reporting/interfaces.ts
export interface IRPTTeam {
  id: string;
  name: string;
  projects_count: number;
  members: InlineMember[];
  selected: boolean;
}
// ...
export interface IRPTOverviewProject {
  id: string;
  name: string;
  client: string;
  status: {
    name: string;
    color_code: string;
    icon: string;
  };
}

export interface IRPTOverviewProjectExt extends IRPTOverviewProject {
  team_member_id: string;
}

export interface IRPTMemberResponse {
  total: number,
  members: IRPTMember[],
  team: {
    id: string,
    name: string,
  }
}

export interface IRPTMember {
  color_code: string;
  tasks_stat: any;

  avatar_url?: string;
  completed: number;
  /** Team member id */
  email: string;
  id: string;
  name: string;
  teams: string;
  projects: number;
  tasks: number;
  overdue: number;
  ongoing: number;
  todo: number;
  member_teams: any;
}
// ...
export interface IRPTMemberDrawerData {
  project: IRPTOverviewProject | null;
  member: IRPTMember | null;
}